+++ src/core/renderer/Renderer.cpp
@@ -164,6 +164,8 @@ namespace YimMenu
 		return ImGui_ImplWin32_WndProcHandler(hwnd, msg, wparam, lparam);
 	}
 
+	//This is needed because of how DX12 resizing works. It waits for the last frame to finish then invalidates objects. We would use fence for that.
+	//	However, it requires alot of work and logic, making it not worth it. In all of my testing, just simply waiting a bit works.
 	void Renderer::WaitForLastFrame()
 	{
 		std::this_thread::sleep_for(200ms);
